# tt/node/swagger.json
{
  "info": {
    "title": "fusion-swagger",
    "version": "1.0.0",
    "description": "internal-server swagger"
  },
  "openapi": "3.0.0",
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "authorization"
      }
    }
  },
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
...
  "paths": {
...
    "/api/auth/authorize": {
      "post": {
        "summary": "Check authorization cookie",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "cookie": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ],
        "responses": {
          "200": {
            "description": "user data"
          },
          "401": {
            "description": "Error message, probably wrong request"
          },
          "404": {
            "description": "error no such user"
          }
        }
      }
    },
    "/api/auth/refresh-token": {
      "post": {
        "summary": "Check refresh token",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "cookie": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "auth refresh-token"
        ],
        "responses": {
          "200": {
            "description": "user data"
          },
          "401": {
            "description": "Error message, probably wrong request"
          }
        }
      }
    },
    "/api/auth/password_restore": {
      "post": {
        "summary": "Request password restore",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ],
        "responses": {
          "200": {
            "description": "email have been sent to the email"
          },
          "403": {
            "description": "Error message, probably wrong request"
          },
          "404": {
            "description": "error no such user"
          }
        }
      }
    },
    "/api/auth/reset/{token}": {
      "post": {
        "summary": "reset password",
        "parameters": [
          {
            "in": "path",
            "name": "token",
            "schema": {
              "type": "string"
            },
            "description": "token generated onRestore"
          },
          {
            "name": "reset pass object"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "newPass": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ],
        "responses": {
          "200": {
            "description": "password have been changed"
          },
          "400": {
            "description": "no token"
          },
          "404": {
            "description": "err message. Probably no such user"
          }
        }
      }
    },
    "/api/auth/signin": {
      "post": {
        "summary": "log in",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "login": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ],
        "responses": {
          "200": {
            "description": "password have been changed"
          },
          "500": {
            "description": "validation errors"
          }
        }
      }
    },
    "/api/auth/signup": {
      "post": {
        "summary": "register",
        "produces": [
          "application/json"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string"
                  },
                  "login": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ],
        "responses": {
          "200": {
            "description": "password have been changed"
          }
        }
      }
    },
...
  },
...
}